WIP: Refactor communties and collections with regards to admin actions
