+++ app/views/admin/collections/_form.html.erb
@@ -1,4 +1,4 @@
-<%= simple_form_for [community, collection] do |f| %>
+<%= simple_form_for [:admin, community, collection] do |f| %>
   <%= f.input :title %>
   <%= f.input :description, as: :text %>
   <%= f.input :community_id, as: :hidden %>
