+++ app/views/admin/collections/edit.html.erb
