+++ app/views/admin/communities/_form.html.erb
@@ -1,0 +1,23 @@
+<%= simple_form_for [:admin, @community] do |f| %>
+  <div class="form-group">
+    <% if @community.logo.attached? %>
+      <%= link_to image_tag(@community.logo.url, class: "img-thumbnail community-logo-small"),
+          @community.logo.url, target: :_blank %>
+    <% else %>
+      <div class="img-thumbnail community-logo-small no-logo">
+        <i class="fa fa-3x fa-picture-o" aria-hidden="true"></i>
+        <p class="community-no-logo-text"><%= t('communities.has_no_logo') %></p>
+      </div>
+    <% end %>
+
+    <%= f.file_field :logo %>
+
+    <% if @community.logo.attached? %>
+      <%= f.input :remove_logo, as: :boolean, checked_value: true, unchecked_value: false %>
+    <% end %>
+
+    <%= f.input :title %>
+    <%= f.input :description, as: :text %>
+    <%= f.button :submit %>
+  </div>
+<% end %>
