+++ app/views/admin/communities/show.js.erb
@@ -1,4 +1,4 @@
-$("ul#<%=@community.id%>").html("<%= escape_javascript(render partial: 'communities/collections_list') %>");
+$("ul#<%=@community.id%>").html("<%= escape_javascript(render partial: 'admin/communities/collections_list') %>");
 $("ul#<%=@community.id%>").show();
 $(".collection-btn[data-community-id='<%=@community.id%>']").hide();
 $(".close-btn[data-community-id='<%=@community.id%>']").removeClass('d-none');
