--- app/views/communities/_form.html.erb
@@ -1,9 +1,0 @@
-<%= simple_form_for @community do |f| %>
-  <div class="form-group">
-    <%= render partial: 'logo' %>
-    <%= render partial: 'logo_form', locals: { f: f } %>
-    <%= f.input :title %>
-    <%= f.input :description, as: :text %>
-    <%= f.button :submit %>
-  </div>
-<% end %>
