--- app/views/communities/_logo.html.erb
@@ -1,9 +1,0 @@
-<% if @community.logo.attached? %>
-  <%= link_to image_tag(@community.logo.url, class: "img-thumbnail community-logo-small"),
-      @community.logo.url, target: :_blank %>
-<% else %>
-  <div class="img-thumbnail community-logo-small no-logo">
-    <i class="fa fa-3x fa-picture-o" aria-hidden="true"></i>
-    <p class="community-no-logo-text"><%= t('.has_no_logo') %></p>
-  </div>
-<% end %>
