--- app/views/communities/_logo_form.html.erb
@@ -1,4 +1,0 @@
-<%= f.file_field :logo %>
-<% if @community.logo.attached? %>
-  <%= f.input :remove_logo, as: :boolean, checked_value: true, unchecked_value: false %>
-<% end %>
