+++ app/views/communities/show.html.erb
@@ -7,11 +7,11 @@
     <div class="col text-right" >
       <div class="btn-group">
         <% if policy(@community).edit? %>
-          <%= link_to t('.edit'), edit_community_path(@community),
+          <%= link_to t('edit'), edit_admin_community_path(@community),
               class:'btn btn-outline-primary' %>
         <% end %>
         <% if policy(@community).destroy? %>
-          <%= link_to t('.delete'), community_path(@community), method: :delete,
+          <%= link_to t('delete'), admin_community_path(@community), method: :delete,
               class:'btn btn-danger',
               data: { confirm: t('.delete_confirm', title: @community.title) } %>
         <% end %>
@@ -20,9 +20,18 @@
   </div>
 <% end %>
 
+
 <div class="d-flex justify-content-between m-3">
   <div class="p-3 align-self-center">
-    <%= render partial: 'logo' %>
+    <% if @community.logo.attached? %>
+      <%= link_to image_tag(@community.logo.url, class: "img-thumbnail community-logo-small"),
+          @community.logo.url, target: :_blank %>
+    <% else %>
+      <div class="img-thumbnail community-logo-small no-logo">
+        <i class="fa fa-3x fa-picture-o" aria-hidden="true"></i>
+        <p class="community-no-logo-text"><%= t('communities.has_no_logo') %></p>
+      </div>
+    <% end %>
   </div>
 
   <div class="p-3 align-self-center">
@@ -37,33 +46,39 @@
     <%= t('.collections_list_header') %>
   </h4>
   <% if policy(:collection).create? %>
-      <%= link_to t('.create_new_collection'), new_community_collection_path(@community),
+      <%= link_to t('.create_new_collection'), new_admin_community_collection_path(@community),
           class: 'btn btn-primary align-self-center' %>
   <% end %>
 </div>
 
 <div class="row mt-3">
   <div class="col-md-12">
-    <%# TODO: FIX ALL THIS... %>
-    <% if policy(:admin).access? %>
-      <ul class="list-group">
-        <%= render partial: 'collections_list' %>
-      </ul>
-    <% else %>
+    <ul class="list-group">
       <% if @community.member_collections.count > 0 %>
-        <div class="list-group">
           <% @community.member_collections.each do |collection| %>
-            <%= link_to collection.title, community_collection_path(@community, collection),
-                class: 'list-group-item list-group-item-action' %>
-          <% end %>
-        </div>
+            <li class="list-group-item list-group-item-action d-flex justify-content-between">
+              <%= link_to collection.title, community_collection_path(@community, collection) %>
+
+              <% if policy(:admin).access? %>
+                <div class="btn-group align-self-center">
+                  <% if policy(collection).edit? %>
+                    <%= link_to t('edit'), edit_admin_community_collection_path(@community, collection),
+                        class:'btn btn-outline-primary edit-collection' %>
+                  <% end %>
+                  <% if policy(collection).destroy? %>
+                    <%= link_to t('delete'), admin_community_collection_path(@community, collection),
+                        class: 'btn btn-danger delete-collection', method: :delete,
+                        data: { confirm: t('collections.show.delete_confirm', title: collection.title) } %>
+                  <% end %>
+                </div>
+              <% end %>
+            <% end %>
+            </li>
       <% else %>
-        <ul class="list-group">
-          <li class="list-group-item list-group-item-action">
-            <%= t('communities.show.no_collections') %>
-          </li>
-        </ul>
+        <li class="list-group-item list-group-item-action">
+          <%= t('communities.show.no_collections') %>
+        </li>
       <% end %>
-    <% end %>
+    </ul>
   </div>
 </div>
