+++ app/views/layouts/admin.html.erb
@@ -18,9 +18,9 @@
         <li class="nav-item"><%= link_to t('admin.users.index.header'), admin_users_path, class: 'nav-link' %></li>
       <% end %>
 
-      <%= render 'sidebar_group', title: t('admin.communities_and_collections.title') do %>
-        <li class="nav-item"><%= link_to t('admin.communities_and_collections.index.header'), admin_communities_and_collections_path, class: 'nav-link' %></li>
-        <li class="nav-item"><%= link_to t('admin.communities_and_collections.new.header'), new_community_path, class: 'nav-link' %></li>
+      <%= render 'sidebar_group', title: t('admin.communities.title') do %>
+        <li class="nav-item"><%= link_to t('admin.communities.index.header'), admin_communities_path, class: 'nav-link' %></li>
+        <li class="nav-item"><%= link_to t('admin.communities.new.header'), new_admin_community_path, class: 'nav-link' %></li>
       <% end %>
 
       <%= render 'sidebar_group', title: t('admin.site_notifications.title') do %>
